requirements.txt
crewai==0.28.8
langchain-google-genai==1.0.3
langchain-community==0.0.38
duckduckgo-search==5.3.0
fpdf2==2.7.8
streamlit==1.34.0
python-dotenv==1.0.1